// src/app/Modules/carrera/model/cpconvocatoria.dto.ts
// Generated by https://quicktype.io

export interface CpConvocatoriaDTO {
    any:               number;
    dataReconeixement: string;
    descripcio:        string;
    estat:             string;
    kconvocatoria:     number;
    tipus:             string;
    estatStr:          string;
    tipusStr:          string;

}