// src/app/Modules/carrera/model/cpresultatparticipacio.dto.ts
// Generated by https://quicktype.io

import { RelPuePersDTO } from "../../rrhh/model/relpuepers.dto";
import { CpAvaluacioDTO } from "./cpavaluacio.dto";
import { CpCursAportatDTO } from "./cpcursaportat.dto";
import { CpParticipacioDTO } from "./cpparticipacio.dto";
import { CpPersonalDTO } from "./cppersonal.dto";
import { CpResultatAvaluacioDTO } from "./cpresultatavaluacio.dto";
import { CpServPrestatDTO } from "./cpservprestat.dto";
import { CpTitolAportatDTO } from "./cptitolaportat.dto";

export interface CpResultatParticipacioDTO {
    cursos:        CpCursAportatDTO[];
    cursosant:     CpCursAportatDTO[];
    ocupacio:      CpServPrestatDTO[];
    participacio:  CpParticipacioDTO;
    personal:      CpPersonalDTO;
    puntsRequrits: number;
    rptactual:     RelPuePersDTO;
    titols:        CpTitolAportatDTO[];
    titolsant:     CpTitolAportatDTO[];
    requisits:     string[];
    resultatavaluacio: CpResultatAvaluacioDTO;
}